Remove magic variables check from critic
Discussing with Derek on the ProPublica issues, there should be
stricter checking for users making API calls.  Although it would
become a headache to update all range values for the calls, I want
to work with some restrictions in the SDK.  Rather than setting
each boundary as a constant, the limits can be defined in the
method itself.  The places being defined should be self descriptive
and not need excessive commenting.